+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Nextcloud 25 compatibility
+- Updated NPM dependencies
+
 ## [1.1.2] - 2022-09-08
 ### Fixed
 - Fixed too long payload column ([#31](https://github.com/mziech/nextcloud-majordomo/issues/31))